global: Don't fail if SDClang variables aren't defined
CAF assumes that SDClang will always be in the environment, but
there are cases where executing soong is done without the variables,
and thus causes startup failures. Workaround this by checking for
SDCLANG_CONFIG to be defined and by checking if SDClang is enabled
before failing due to SDCLANG_PATH being empty.

Change-Id: I193815fb957d8830f0e39f1e8dd9b53eac47b60c
Signed-off-by: hpnightowl <iamhp2k@gmail.com>